Bug 1008473 - Remove some redundant methods and misc tidyup r=terrence
UltraBlame original commit: bff5585f1b9b47a2fc424f916e295a2fda81880a

+++ js/public/HeapAPI.h
@@ -58,8 +58,8 @@ static const uint32_t GRAY = 1;
 
 
 
-const uintptr_t ChunkLocationNursery = 0;
-const uintptr_t ChunkLocationTenuredHeap = 1;
+const uint32_t ChunkLocationNursery = 0;
+const uint32_t ChunkLocationTenuredHeap = 1;
 
 #ifdef JS_DEBUG
